# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: СМП (Система медленных платежей) API
  version: 1.0.0
  description: |
    Единая спецификация API-контрактов для взаимодействия микросервисов СМП.
servers:
  - url: http://localhost:8080
    description: Gateway (единственная точка входа для внешних клиентов)
  - url: http://localhost:8081
    description: Switch Service
  - url: http://localhost:8082
    description: Authorization Service
  - url: http://localhost:8083
    description: Card Management Service
  - url: http://localhost:8084
    description: Terminal Simulator
  - url: http://localhost:8085
    description: Merchant Simulator
  - url: http://localhost:8088
    description: Transaction Logger
...
paths:
...
  /api/cards:
    post:
      tags: [Cards]
      summary: Создать новую банковскую карту
      description:
        Создаёт карту с автоматической генерацией PAN по алгоритму Луна.
        Срок действия = текущая дата + 3 года. Начальный статус - ACTIVE.
      operationId: createCard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardCreateRequest'
      responses:
        '201':
          description: Карта успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '400':
          description: Ошибка валидации (некорректный BIN, имя и т.п.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags: [Cards]
      summary: Получить список карт с пагинацией и фильтрами
      description: Основной эндпоинт для получения пула карт. Карты со статусом DELETED не возвращаются.
      operationId: listCards
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
          description: Количество карт на странице
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Смещение от начала выборки
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [ACTIVE, INACTIVE, BLOCKED, EXPIRED]
          description: Фильтр по статусу карты
        - name: bin
          in: query
          required: false
          schema:
            type: string
            pattern: '^\d{6}$'
          description: Фильтр по BIN (первые 6 цифр PAN)
        - name: issuerId
          in: query
          required: false
          schema:
            type: string
            pattern: '^[A-Z0-9]+$'
          description: Идентификатор банка эмитента
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            pattern: date-time
          description: Минимальная дата создания
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            pattern: date-time
          description: Максимальная дата создания
      responses:
        '200':
          description: Список карт
          content:
            application/json:
              schema:
                type: object
                required: [total, cards]
                properties:
                  total:
                    type: integer
                    description: Общее количество карт, удовлетворяющих фильтрам
                    example: 100
                  cards:
                    type: array
                    items:
                      $ref: '#/components/schemas/Card'
        '400':
          description: Ошибка валидации (некорректный BIN, имя и т.п.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cards/generate:
    post:
      tags: [Cards]
      summary: Массовая генерация тестовых карт
      description: |
        Генерирует указанное количество тестовых карт, равномерно распределённых по заданным BIN.
        Используется симуляторами и операторами для наполнения системы тестовыми данными
        перед запуском сценариев. Для каждой карты генерируется случайное имя держателя,
        баланс (10k–500k ₽), дневной лимит (50k–300k ₽) и месячный лимит (дневной × 30).
        Распределение статусов: ACTIVE 95%, INACTIVE 3%, BLOCKED 2%.
      operationId: generateCards
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateCardsRequest'
      responses:
        '201':
          description: Карты успешно сгенерированы
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateCardsResponse'
        '400':
          description: Ошибка валидации (пустой bins, count = 0 и т.п.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/simulator/merchant/run:
    post:
      tags: [Simulator]
      summary: Запустить симуляцию мерчантов и эквайреров
      description: |
        Генерирует транзакции с разнообразными MCC-кодами, комиссиями эквайрера
        и типами бизнеса. В отличие от Terminal Simulator, добавляет реалистичности
        за счёт разных категорий мерчантов (супермаркеты, рестораны, электроника,
        e-commerce, телеком, отели и т.д.).
        Доступные сценарии:
        - `grocery` - продуктовые магазины (MCC 5411, 5499, суммы 100-3000 ₽)
        - `electronics` - электроника (MCC 5732, 5045, суммы 5 000-100 000 ₽)
        - `restaurant` - рестораны (MCC 5812, 5814, суммы 500-5000 ₽)
        - `travel` - путешествия (MCC 3501, 4511, 4722, суммы 5 000-50 000 ₽)
      operationId: runMerchantSimulator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MerchantSimulatorRequest'
      responses:
        '200':
          description: Симуляция завершена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulatorResponse'
...
  /api/dashboard/recent:
    get:
      tags: [Dashboard]
      summary: Получить последние N транзакций для таблицы дашборда
      description:
        Возвращает список самых свежих транзакций.
      operationId: getRecentTransactions
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: Количество транзакций
      responses:
        '200':
          description: Список последних транзакций
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
  /api/dashboard/charts:
    get:
      tags: [Dashboard]
      summary: Получить агрегацию транзакций по времени для графиков
      description: Возвращает почасовую или подневную агрегацию для построения графиков на дашборде.
      operationId: getDashboardCharts
      parameters:
        - name: granularity
          in: query
          required: false
          schema:
            type: string
            enum: [hour, day]
            default: hour
          description: Гранулярность агрегации
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Начало диапазона
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Конец диапазона
      responses:
        '200':
          description: Временные ряды
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required: [timestamp, total, approved, declined, amount]
                  properties:
                    timestamp:
                      type: string
                      format: date-time
                    total:
                      type: integer
                    approved:
                      type: integer
                    declined:
                      type: integer
                    amount:
                      type: integer
                      description: Сумма транзакций в копейках
...
components:
  schemas:
    ErrorResponse:
      type: object
      required: [error, message, timestamp]
      properties:
        error:
          type: string
          description: Машинно-читаемый код ошибки
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Человеко-читаемое описание ошибки
          example: "Field 'pan' must be exactly 16 digits"
        timestamp:
          type: string
          format: date-time
          description: Время возникновения ошибки
        serviceName:
          type: string
          description: Имя недоступного сервиса
        retryAfterMs:
          type: integer
          description: Через сколько миллисекунд повторить запрос
...
    CardCreateRequest:
      type: object
      required: [bin, cardholderName]
      properties:
        bin:
          type: string
          pattern: '^\d{6}$'
          description: Bank Identification Number - первые 6 цифр PAN
          example: "400000"
        cardholderName:
          type: string
          maxLength: 255
          description: Имя и фамилия держателя карты латиницей (как на карте)
          example: "IVAN IVANOV"
        currencyCode:
          type: string
          pattern: '^\d{3}$'
          default: "643"
          description: Код валюты ISO 4217 (643 = RUB, 840 = USD, 978 = EUR)
        dailyLimit:
          type: integer
          default: 15000000
          description: Дневной лимит расходов в копейках (150 000 ₽ по умолчанию)
        monthlyLimit:
          type: integer
          default: 300000000
          description: Месячный лимит расходов в копейках (3 000 000 ₽ по умолчанию)
        initialBalance:
          type: integer
          default: 100000000
          description: Начальный баланс карты в копейках (1 000 000 ₽ по умолчанию)
    Card:
      type: object
      required:
        - id
        - pan
        - bin
        - cardholderName
        - expiryDate
        - status
        - currencyCode
        - dailyLimit
        - monthlyLimit
        - availableBalance
        - issuerId
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Уникальный идентификатор карты в системе
        pan:
          type: string
          pattern: '^\d{16}$'
          description: Primary Account Number - полный номер карты (16 цифр, проходит проверку Луна)
          example: "4000001234560001"
        bin:
          type: string
          pattern: '^\d{6}$'
          description: Bank Identification Number - первые 6 цифр PAN
        cardholderName:
          type: string
          description: Имя держателя карты
        expiryDate:
          type: string
          pattern: '^\d{4}$'
          description: Срок действия в формате MMYY
          example: "1229"
        status:
          type: string
          enum: [ACTIVE, INACTIVE, BLOCKED, EXPIRED, DELETED]
          description: |
            Статус карты:
            * `ACTIVE` - активна, готова к использованию
            * `INACTIVE` - неактивна (например, сгенерирована как неактивная)
            * `BLOCKED` - заблокирована
            * `EXPIRED` - срок действия истёк
            * `DELETED` - мягко удалена (не возвращается в API)
        currencyCode:
          type: string
          pattern: '^\d{3}$'
          description: Код валюты ISO 4217
        dailyLimit:
          type: integer
          description: Дневной лимит расходов в копейках
        monthlyLimit:
          type: integer
          description: Месячный лимит расходов в копейках
        availableBalance:
          type: integer
          description: Доступный остаток на карте в копейках (уменьшается при резервировании)
        issuerId:
          type: string
          description: ID банка-эмитента (определяется по BIN)
          example: "ISS001"
        createdAt:
          type: string
          format: date-time
          description: Дата и время создания карты
...
    GenerateCardsRequest:
      type: object
      required: [count, bins]
      properties:
        count:
          type: integer
          minimum: 1
          maximum: 10000
          description: Количество карт для генерации
          example: 100
        bins:
          type: array
          items:
            type: string
            pattern: '^\d{6}$'
          minItems: 1
          description: Список BIN-префиксов. Карты распределяются равномерно между ними.
          example: ["400000", "400001", "400002"]
    GenerateCardsResponse:
      type: object
      required: [generated, cards]
      properties:
        generated:
          type: integer
          description: Фактическое количество сгенерированных карт
        cards:
          type: array
          items:
            $ref: '#/components/schemas/Card'
...
    AuthorizationRequest:
      type: object
      required:
        - mti
        - stan
        - pan
        - processingCode
        - amount
        - currencyCode
        - transmissionDateTime
        - terminalId
        - merchantId
        - mcc
        - acquirerId
      properties:
        mti:
          type: string
          pattern: '^\d{4}$'
          description: |
            Message Type Identifier - тип сообщения ISO 8583
          example: "0100"
        stan:
          type: string
          pattern: '^\d{6}$'
          description: System Trace Audit Number - уникальный номер транзакции в терминале (000001..999999)
          example: "000001"
        pan:
          type: string
          pattern: '^\d{16}$'
          description: Primary Account Number - номер карты (16 цифр)
          example: "4000001234560001"
        processingCode:
          type: string
          pattern: '^\d{6}$'
          description: Код обработки (000000 = покупка, 000001 = снятие наличных)
          example: "000000"
        amount:
          type: integer
          minimum: 1
          description: Сумма транзакции в минорных единицах (копейках). 150000 = 1 500.00 ₽
          example: 150000
        currencyCode:
          type: string
          pattern: '^\d{3}$'
          description: Код валюты ISO 4217 (643 = RUB)
          example: "643"
        transmissionDateTime:
          type: string
          format: date-time
          description: Дата и время отправки транзакции терминалом (ISO 8601, UTC)
        terminalId:
          type: string
          maxLength: 8
          description: Идентификатор POS-терминала
          example: "TERM001"
        terminalType:
          type: string
          enum: [POS, ATM, ECOM]
          description: Тип терминала (опционально)
        merchantId:
          type: string
          maxLength: 15
          description: Идентификатор мерчанта (торговой точки)
          example: "MERCH12345678901"
        mcc:
          type: string
          pattern: '^\d{4}$'
          description: Merchant Category Code - код категории бизнеса
          example: "5411"
        acquirerId:
          type: string
          description: ID банка-эквайера
          example: "ACQ001"
        issuerId:
          type: string
          description: |
            ID банка-эмитента. Не заполняется клиентом - определяется Switch-ом
            по BIN карты на этапе маршрутизации.
          example: "ISS001"
    AuthorizationResponse:
      type: object
      required:
        - mti
        - stan
        - rrn
        - authCode
        - responseCode
        - status
      properties:
        mti:
          type: string
          example: "0110"
          description: Message Type Identifier ответа (0110 = ответ на авторизацию 0100)
        stan:
          type: string
          description: STAN из исходного запроса (для сопоставления)
        rrn:
          type: string
          maxLength: 12
          description: |
            Retrieval Reference Number - уникальный идентификатор транзакции в системе.
            Формат: `{YDDD}{HH}{mm}{ss}{seq}`. Уникален в рамках всей системы.
          example: "012345678901"
        authCode:
          type: string
          maxLength: 6
          description: Код авторизации (6 символов A-Z0-9). Генерируется только при APPROVED.
          example: "ABC123"
        responseCode:
          type: string
          pattern: '^\d{2}$'
          description: |
            Код ответа ISO 8583:
            * `00` - Approved (одобрено)
            * `05` - Do Not Honor (отказано, общая причина)
            * `14` - Invalid Card Number (карта не найдена / неверный BIN)
            * `51` - Insufficient Funds (недостаточно средств)
            * `54` - Expired Card (карта просрочена)
            * `61` - Exceeds Amount Limit (превышен дневной/месячный лимит)
            * `96` - System Error (системная ошибка, например, откат из-за недоступности Logger)
          example: "00"
        status:
          type: string
          enum: [APPROVED, DECLINED]
          description: Итоговый статус транзакции
        declineReason:
          type: string
          nullable: true
          description: |
            Человеко-читаемая причина отклонения. Заполняется только при status=DECLINED.
            Примеры: "Insufficient funds", "CARD_BLOCKED", "ISSUER_TIMEOUT".
          example: "Insufficient funds"
        processingTimeMs:
          type: integer
          description: Время обработки в Authorization Service, мс.
          example: 42
    SimulatorResponse:
      type: object
      required: [totalSubmitted, approved, declined, elapsedMs]
      properties:
        totalSubmitted:
          type: integer
          description: Общее количество отправленных транзакций
        approved:
          type: integer
          description: Количество одобренных транзакций
        declined:
          type: integer
          description: Количество отклонённых транзакций
        elapsedMs:
          type: integer
          description: Общее время выполнения симуляции в миллисекундах
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/AuthorizationResponse'
          description: Детализация ответов по каждой транзакции
...
    MerchantSimulatorRequest:
      type: object
      required: [count, scenario]
      properties:
        count:
          type: integer
          minimum: 1
          description: Количество транзакций
          example: 50
        scenario:
          type: string
          enum: [grocery, electronics, restaurant, travel]
          description: Сценарий симуляции
        mccCodes:
          type: array
          items:
            type: string
            pattern: '^\d{4}$'
          description: |
            Явный список MCC-кодов. Если указан — используется вместо MCC по умолчанию
            для выбранного сценария.
    Transaction:
      type: object
      description: |
        Полная запись о транзакции в журнале. Объединяет AuthorizationRequest +
        AuthorizationResponse + системные поля (id, acquiringFee, createdAt).
      allOf:
        - $ref: '#/components/schemas/AuthorizationRequest'
        - $ref: '#/components/schemas/AuthorizationResponse'
        - type: object
          required: [id, createdAt]
          properties:
            id:
              type: string
              format: uuid
              description: Уникальный идентификатор записи транзакции в БД Logger
            acquiringFee:
              type: integer
              nullable: true
              description: Комиссия эквайера в копейках (рассчитывается Merchant Simulator)
            createdAt:
              type: string
              format: date-time
              description: Дата и время записи транзакции в Logger
...